Encapsulate UI module in ui namespace
- Also put compile options in top CMakeLists.txt
